chore: Update Dockerfile and requirements.txt
This commit updates the Dockerfile to improve the debug information and run the Flask app using JSON format. It also updates the requirements.txt file with the latest dependencies.

```
