+++ requirements.txt
@@ -1,99 +1,90 @@
-annotated-types==0.7.0
-anthropic==0.34.0
-anyio==4.4.0
-blinker==1.8.2
-cachetools==5.4.0
-certifi==2024.7.4
-cffi==1.17.0
-charset-normalizer==3.3.2
-click==8.1.7
-cryptography==43.0.0
-deprecated==1.2.14
-distro==1.9.0
-dnspython==2.6.1
-fastapi==0.112.1
-filelock==3.15.4
-flask==3.0.3
-fsspec==2024.6.1
-google-api-core==2.19.1
-google-api-python-client==2.141.0
-google-auth==2.33.0
-google-auth-httplib2==0.2.0
-googleapis-common-protos==1.63.2
-groq==0.9.0
-h11==0.14.0
-httpcore==1.0.5
-httplib2==0.22.0
-httpx==0.27.0
-huggingface-hub==0.24.5
-idna==3.7
-itsdangerous==2.2.0
-jinja2==3.1.4
-jiter==0.5.0
-joblib==1.4.2
-loguru==0.7.2
-markdown-it-py==3.0.0
-markupsafe==2.1.5
-mdurl==0.1.2
-mpmath==1.3.0
-networkx==2.8.8
-nltk==3.9.1
-numpy==1.26.4
-nvidia-cublas-cu12==12.1.3.1
-nvidia-cuda-cupti-cu12==12.1.105
-nvidia-cuda-nvrtc-cu12==12.1.105
-nvidia-cuda-runtime-cu12==12.1.105
-nvidia-cudnn-cu12==9.1.0.70
-nvidia-cufft-cu12==11.0.2.54
-nvidia-curand-cu12==10.3.2.106
-nvidia-cusolver-cu12==11.4.5.107
-nvidia-cusparse-cu12==12.1.0.106
-nvidia-nccl-cu12==2.20.5
-nvidia-nvjitlink-cu12==12.6.20
-nvidia-nvtx-cu12==12.1.105
-packaging==24.1
-pillow==10.4.0
-pinecone-client==5.0.1
-pinecone-plugin-inference==1.0.3
-pinecone-plugin-interface==0.0.7
-proto-plus==1.24.0
-protobuf==5.27.3
-pyasn1==0.6.0
-pyasn1-modules==0.4.0
-pycparser==2.22
-pydantic==2.8.2
-pydantic-core==2.20.1
-pygithub==2.3.0
-pygments==2.18.0
-pyjwt[crypto]==2.9.0
-pymongo==4.8.0
-pynacl==1.5.0
-pyparsing==3.1.2
-python-dateutil==2.9.0.post0
-python-dotenv==1.0.1
-pyyaml==6.0.2
-redis==5.0.8
-regex==2024.7.24
-requests==2.32.3
-rich==13.7.1
-rsa==4.9
-safetensors==0.4.4
-scikit-learn==1.5.1
-scipy==1.11.4
-sentence-transformers==3.0.1
-six==1.16.0
-sniffio==1.3.1
-starlette==0.38.2
-sympy==1.13.2
-threadpoolctl==3.5.0
-tokenizers==0.19.1
-torch==2.4.0
-tqdm==4.66.5
-transformers==4.44.0
-triton==3.0.0
-typing-extensions==4.12.2
-uritemplate==4.1.1
-urllib3==2.2.2
-uvicorn==0.30.6
-werkzeug==3.0.3
-wrapt==1.16.0
+annotated-types
+anthropic
+anyio
+blinker
+cachetools
+certifi
+cffi
+charset-normalizer
+click
+cryptography
+deprecated
+distro
+dnspython
+fastapi
+filelock
+flask
+fsspec
+google-api-core
+google-api-python-client
+google-auth
+google-auth-httplib2
+googleapis-common-protos
+groq
+h11
+httpcore
+httplib2
+httpx
+huggingface-hub
+idna
+itsdangerous
+jinja2
+jiter
+joblib
+langchain
+langchain-community
+loguru
+markdown-it-py
+markupsafe
+mdurl
+mpmath
+networkx
+nltk
+numpy
+openai
+packaging
+pillow
+pinecone-client
+pinecone-plugin-inference
+pinecone-plugin-interface
+proto-plus
+protobuf
+pyasn1
+pyasn1-modules
+pycparser
+pydantic
+pydantic-core
+pygithub
+pygments
+pyjwt[crypto]
+pymongo
+pynacl
+pyparsing
+python-dateutil
+python-dotenv
+pyyaml
+redis
+regex
+requests
+rich
+rsa
+safetensors
+scikit-learn
+scipy
+sentence-transformers
+six
+sniffio
+starlette
+sympy
+threadpoolctl
+tokenizers
+torch
+tqdm
+transformers
+triton
+typing-extensions
+uritemplate
+urllib3
+uvicorn
+werkzeug
+wrapt
